+++ examples/widgetbook_annotation_example/lib/app.widgetbook.dart
@@ -5,21 +5,13 @@
 // **************************************************************************
 
 import 'dart:core';
+
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:meal_app/app.dart';
-import 'package:meal_app/constants/border.dart';
-import 'package:meal_app/constants/color.dart';
-import 'package:meal_app/models/meal.dart';
 import 'package:meal_app/themes/dark_theme.dart';
 import 'package:meal_app/themes/light_theme.dart';
 import 'package:meal_app/widgets/attributes/attribute.dart';
-import 'package:meal_app/widgets/attributes/price_attribute.dart';
-import 'package:meal_app/widgets/attributes/weight_attribute.dart';
-import 'package:meal_app/widgets/ingredients.dart';
 import 'package:meal_app/widgets/meal_detail.dart';
-import 'package:meal_app/widgets/new_tag.dart';
-import 'package:meal_app/widgets/rotated_image.dart';
 import 'package:widgetbook/widgetbook.dart';
 
 void main() {
@@ -35,45 +27,106 @@ class HotReload extends StatelessWidget {
       appInfo: AppInfo(
         name: 'Meal App',
       ),
-      supportedLocales: locales,
-      localizationsDelegates: delegates,
-      themes: [
-        WidgetbookTheme(
-          name: 'Dark',
-          data: getDarkThemeData(),
-        ),
-        WidgetbookTheme(
-          name: 'Light',
-          data: getLightThemeData(),
-        ),
-      ],
-      devices: [
-        Device(
-          name: 'iPhone 12',
-          resolution: Resolution(
-            nativeSize: DeviceSize(
-              height: 2532.0,
-              width: 1170.0,
+      addons: [
+        CustomThemeAddon<ThemeData>(
+          setting: ThemeSetting<ThemeData>(
+            themes: [
+              WidgetbookTheme(
+                name: 'Dark',
+                data: getDarkThemeData(),
+              ),
+              WidgetbookTheme(
+                name: 'Light',
+                data: getLightThemeData(),
+              ),
+            ],
+            activeTheme: WidgetbookTheme(
+              name: 'Dark',
+              data: getDarkThemeData(),
             ),
-            scaleFactor: 3.0,
           ),
-          type: DeviceType.mobile,
         ),
-      ],
-      frames: [
-        WidgetbookFrame(
-          name: 'Widgetbook',
-          allowsDevices: true,
+        TextScaleAddon(
+          setting: TextScaleSetting(
+            textScales: [
+              1.0,
+              2.0,
+              3.0,
+            ],
+            activeTextScale: 1.0,
+          ),
         ),
-        WidgetbookFrame(
-          name: 'None',
-          allowsDevices: false,
+        LocalizationAddon(
+          setting: LocalizationSetting(
+            locales: locales,
+            activeLocale: locales.first,
+            localizationsDelegates: delegates,
+          ),
         ),
-      ],
-      textScaleFactors: [
-        1.0,
-        2.0,
-        3.0,
+        FrameAddon(
+          setting: FrameSetting(
+            frames: [
+              NoFrame(),
+              DefaultDeviceFrame(
+                setting: DeviceSetting(
+                  devices: [
+                    Device(
+                      name: 'iPhone 12',
+                      resolution: Resolution(
+                        nativeSize: DeviceSize(
+                          height: 2532.0,
+                          width: 1170.0,
+                        ),
+                        scaleFactor: 3.0,
+                      ),
+                      type: DeviceType.mobile,
+                    ),
+                  ],
+                  activeDevice: Device(
+                    name: 'iPhone 12',
+                    resolution: Resolution(
+                      nativeSize: DeviceSize(
+                        height: 2532.0,
+                        width: 1170.0,
+                      ),
+                      scaleFactor: 3.0,
+                    ),
+                    type: DeviceType.mobile,
+                  ),
+                ),
+              ),
+              WidgetbookFrame(
+                setting: DeviceSetting(
+                  devices: [
+                    Device(
+                      name: 'iPhone 12',
+                      resolution: Resolution(
+                        nativeSize: DeviceSize(
+                          height: 2532.0,
+                          width: 1170.0,
+                        ),
+                        scaleFactor: 3.0,
+                      ),
+                      type: DeviceType.mobile,
+                    ),
+                  ],
+                  activeDevice: Device(
+                    name: 'iPhone 12',
+                    resolution: Resolution(
+                      nativeSize: DeviceSize(
+                        height: 2532.0,
+                        width: 1170.0,
+                      ),
+                      scaleFactor: 3.0,
+                    ),
+                    type: DeviceType.mobile,
+                  ),
+                ),
+              ),
+            ],
+            activeFrame: NoFrame(),
+          ),
+        ),
       ],
       categories: [
         WidgetbookCategory(
@@ -122,10 +175,6 @@ class HotReload extends StatelessWidget {
           widgets: [],
         ),
       ],
-      deviceFrameBuilder: frameBuilder,
-      localizationBuilder: localizationBuilder,
-      scaffoldBuilder: scaffoldBuilder,
-      themeBuilder: themeBuilder(),
     );
   }
 }