+++ packages/widgetbook_generator/lib/code_generators/instances/addons/material_theme_addon/material_theme_addon.dart
@@ -1,0 +1,4 @@
+// TODO this whole implementation is currently not used and is covered by
+// CustomThemeAddon
+export './material_theme_addon_instance.dart';
+export './material_theme_setting_instance.dart';