+++ packages/widgetbook_generator/test/src/code_generators/instances/addons/theme_addon/theme_setting_instance_test.dart
@@ -1,0 +1,22 @@
+import 'package:test/test.dart';
+import 'package:widgetbook_generator/code_generators/instances/instances.dart';
+
+import '../../../../mocks/mocks.dart';
+
+void main() {
+  group('$ThemeSettingInstance', () {
+    test(
+      '.name returns "ThemeSetting"',
+      () {
+        expect(
+          ThemeSettingInstance(
+            themes: themes,
+            defaultTheme: themeLight,
+            themeType: 'ThemeData',
+          ).name,
+          equals('ThemeSetting'),
+        );
+      },
+    );
+  });
+}